+++ packages/core/src/Machine.ts
@@ -6,19 +6,19 @@ import {
   StateSchema,
   EventObject,
   AnyEventObject,
-  Typestate
+  Typestate,
+  EventFrom
 } from './types';
 import { StateNode } from './StateNode';
-import { Model, ModelContextFrom, ModelEventsFrom } from './model';
+import { Model, ModelContextFrom } from './model.types';
 
 /**
  * Docs: {@link https://xstate.js.org/docs/guides/machines.html#configuration Machines}.
  * API: {@link https://xstate.js.org/api/globals.html#machine Machine}.
  *
- * Recommendation: Prefer `createMachine`.
- *
  * Creates a new finite-state machine (statechart) from the config.
  *
+ * @deprecated Use `createMachine(...)` instead.
  */
 export function Machine<
   TContext = any,
@@ -46,15 +46,10 @@ export function Machine<
   options?: Partial<MachineOptions<TContext, TEvent>>,
   initialContext: TContext | (() => TContext) | undefined = config.context
 ): StateMachine<TContext, TStateSchema, TEvent> {
-  const resolvedInitialContext =
-    typeof initialContext === 'function'
-      ? (initialContext as () => TContext)()
-      : initialContext;
-
   return new StateNode<TContext, TStateSchema, TEvent>(
     config,
     options,
-    resolvedInitialContext
+    initialContext
   ) as StateMachine<TContext, TStateSchema, TEvent>;
 }
 
@@ -70,10 +65,10 @@ export function Machine<
 export function createMachine<
   TModel extends Model<any, any, any>,
   TContext = ModelContextFrom<TModel>,
-  TEvent extends EventObject = ModelEventsFrom<TModel>,
+  TEvent extends EventObject = EventFrom<TModel>,
   TTypestate extends Typestate<TContext> = { value: any; context: TContext }
 >(
-  config: MachineConfig<TContext, any, TEvent>,
+  config: MachineConfig<TContext, any, TEvent> & { context: TContext },
   options?: Partial<MachineOptions<TContext, TEvent>>
 ): StateMachine<TContext, any, TEvent, TTypestate>;
 export function createMachine<
@@ -81,7 +76,11 @@ export function createMachine<
   TEvent extends EventObject = AnyEventObject,
   TTypestate extends Typestate<TContext> = { value: any; context: TContext }
 >(
-  config: MachineConfig<TContext, any, TEvent>,
+  // Ensure that only the first overload matches models, and prevent
+  // accidental inference of the model as the `TContext` (which leads to cryptic errors)
+  config: TContext extends Model<any, any, any>
+    ? never
+    : MachineConfig<TContext, any, TEvent>,
   options?: Partial<MachineOptions<TContext, TEvent>>
 ): StateMachine<TContext, any, TEvent, TTypestate>;
 export function createMachine<
@@ -92,14 +91,8 @@ export function createMachine<
   config: MachineConfig<TContext, any, TEvent>,
   options?: Partial<MachineOptions<TContext, TEvent>>
 ): StateMachine<TContext, any, TEvent, TTypestate> {
-  const resolvedInitialContext =
-    typeof config.context === 'function'
-      ? (config.context as () => TContext)()
-      : config.context;
-
   return new StateNode<TContext, any, TEvent, TTypestate>(
     config,
-    options,
-    resolvedInitialContext
+    options
   ) as StateMachine<TContext, any, TEvent, TTypestate>;
 }