+++ packages/core/src/types.ts
@@ -1,6 +1,7 @@
 import { StateNode } from './StateNode';
 import { State } from './State';
 import { Interpreter, Clock } from './interpreter';
+import { Model } from './model.types';
 
 export type EventType = string;
 export type ActionType = string;
@@ -242,10 +243,13 @@ export type Receiver<TEvent extends EventObject> = (
   listener: (event: TEvent) => void
 ) => void;
 
-export type InvokeCallback<TEvent extends EventObject = AnyEventObject> = (
-  callback: Sender<TEvent>,
+export type InvokeCallback<
+  TEvent extends EventObject = AnyEventObject,
+  TSentEvent extends EventObject = AnyEventObject
+> = (
+  callback: Sender<TSentEvent>,
   onReceive: Receiver<TEvent>
-) => any;
+) => (() => void) | Promise<any> | void;
 
 export interface InvokeMeta {
   data: any;
@@ -267,7 +271,7 @@ export interface InvokeMeta {
  */
 export type InvokeCreator<
   TContext,
-  TEvent extends EventObject = AnyEventObject,
+  TEvent extends EventObject,
   TFinalContext = any
 > = (
   context: TContext,
@@ -277,7 +281,8 @@ export type InvokeCreator<
   | PromiseLike<TFinalContext>
   | StateMachine<TFinalContext, any, any>
   | Subscribable<EventObject>
-  | InvokeCallback<TEvent>;
+  | InvokeCallback<any, TEvent>
+  | Behavior<any>;
 
 export interface InvokeDefinition<TContext, TEvent extends EventObject>
   extends ActivityDefinition<TContext, TEvent> {
@@ -611,6 +616,13 @@ export interface StateNodeConfig<
    * The tags for this state node, which are accumulated into the `state.tags` property.
    */
   tags?: SingleOrArray<string>;
+  /**
+   * Whether actions should be called in order.
+   * When `false` (default), `assign(...)` actions are prioritized before other actions.
+   *
+   * @default false
+   */
+  preserveActionOrder?: boolean;
 }
 
 export interface StateNodeDefinition<
@@ -1296,49 +1308,60 @@ export interface Subscription {
 }
 
 export interface Subscribable<T> {
-  subscribe(observer: Observer<T>): Subscription;
   subscribe(
     next: (value: T) => void,
     error?: (error: any) => void,
     complete?: () => void
   ): Subscription;
+  subscribe(observer: Observer<T>): Subscription;
 }
 
 export type Spawnable =
   | StateMachine<any, any, any>
-  | Promise<any>
+  | PromiseLike<any>
   | InvokeCallback
-  | Subscribable<any>;
+  | Subscribable<any>
+  | Behavior<any>;
 
 export type ExtractEvent<
   TEvent extends EventObject,
   TEventType extends TEvent['type']
 > = TEvent extends { type: TEventType } ? TEvent : never;
 
-export interface ActorRef<TEvent extends EventObject, TEmitted = any>
-  extends Subscribable<TEmitted> {
-  send: Sender<TEvent>;
+export interface BaseActorRef<TEvent extends EventObject> {
+  send: (event: TEvent) => void;
 }
 
-export interface SpawnedActorRef<TEvent extends EventObject, TEmitted = any>
-  extends ActorRef<TEvent, TEmitted> {
+export interface ActorRef<TEvent extends EventObject, TEmitted = any>
+  extends Subscribable<TEmitted> {
+  send: Sender<TEvent>; // TODO: this should just be TEvent
   id: string;
   getSnapshot: () => TEmitted | undefined;
   stop?: () => void;
   toJSON?: () => any;
 }
 
+/**
+ * @deprecated Use `ActorRef` instead.
+ */
+export type SpawnedActorRef<
+  TEvent extends EventObject,
+  TEmitted = any
+> = ActorRef<TEvent, TEmitted>;
+
 export type ActorRefFrom<
-  T extends StateMachine<any, any, any> | Promise<any>
+  T extends StateMachine<any, any, any> | Promise<any> | Behavior<any>
 > = T extends StateMachine<infer TContext, any, infer TEvent, infer TTypestate>
-  ? SpawnedActorRef<TEvent, State<TContext, TEvent, any, TTypestate>> & {
+  ? ActorRef<TEvent, State<TContext, TEvent, any, TTypestate>> & {
       /**
-       * @deprecated
+       * @deprecated Use `.getSnapshot()` instead.
        */
       state: State<TContext, TEvent, any, TTypestate>;
     }
   : T extends Promise<infer U>
-  ? SpawnedActorRef<never, U>
+  ? ActorRef<never, U>
+  : T extends Behavior<infer TEvent1, infer TEmitted>
+  ? ActorRef<TEvent1, TEmitted>
   : never;
 
 export type AnyInterpreter = Interpreter<any, any, any, any>;
@@ -1352,4 +1375,46 @@ export type InterpreterFrom<
   infer TTypestate
 >
   ? Interpreter<TContext, TStateSchema, TEvent, TTypestate>
+  : never;
+
+export interface ActorContext<TEvent extends EventObject, TEmitted> {
+  parent?: ActorRef<any, any>;
+  self: ActorRef<TEvent, TEmitted>;
+  id: string;
+  observers: Set<Observer<TEmitted>>;
+}
+
+export interface Behavior<TEvent extends EventObject, TEmitted = any> {
+  transition: (
+    state: TEmitted,
+    event: TEvent,
+    actorCtx: ActorContext<TEvent, TEmitted>
+  ) => TEmitted;
+  initialState: TEmitted;
+  start?: (actorCtx: ActorContext<TEvent, TEmitted>) => TEmitted;
+}
+
+export type EventFrom<T> = T extends StateMachine<any, any, infer TEvent, any>
+  ? TEvent
+  : T extends Model<any, infer TEvent, any>
+  ? TEvent
+  : T extends State<any, infer TEvent, any, any>
+  ? TEvent
+  : T extends Interpreter<any, any, infer TEvent, any>
+  ? TEvent
+  : never;
+
+export type ContextFrom<T> = T extends StateMachine<
+  infer TContext,
+  any,
+  any,
+  any
+>
+  ? TContext
+  : T extends Model<infer TContext, any, any>
+  ? TContext
+  : T extends State<infer TContext, any, any, any>
+  ? TContext
+  : T extends Interpreter<infer TContext, any, any, any>
+  ? TContext
   : never;